Introduce pre-commit hook script
When working on large code bases touching dozens of files that have a
copyright marker in them, taking care of updating the respective
copyright years with each modification requiring so is cumbersome and
easy to forget. However, it also can be automated with relatively simple
means.
One viable approach is an integration with the source code management
system. To that end, this change introduces a script that can be used as
a git pre-commit hook. When installed in a repository, the script will
be invoked on each commit, automatically normalizing any copyright
headers it detects. As an aside, users can circumvent the execution of
the hook by supplying the --no-verify option to git's commit command.