Add support for handling unstaged data
With the first version of the pre-commit git hook script we simply
normalized the files to commit in-place and staged the result
afterwards. This approach is flawed once one does not want to commit all
changes of a file. In such a case, the hook script would falsely stage
the entire content after normalization.
This patch fixes the issue by employing a more sophisticated workflow:
Instead of normalizing the on-disk file content, we first ask git for
the file as it would look like without any unstaged changes and
normalize that. We write out this data into the original file. This
original was previously backed up into a temporary location. We can now
stage the new file for commit. Next, we take the original file we backed
up earlier, normalize it as well, and once more overwrite the file we
worked on. This workflow results in the originally unstaged file
contents not being staged while we still have them on disk as before.