Silently ignore binary files
Copyright headers are usually only contained in true text files. Binary
files, on the other hand, should just be ignored.
This change causes the git pre-commit hook script to silently ignore
unicode conversion errors as they are raised when a byte or sequence of
bytes is found in a file that is not valid in a normal text file.