Print file name in case of error
The git pre-commit might fail in case a copyright year is not parsable,
for example. In such cases we have no information where a problem
occurred other than what the exception itself contains (which might not
be much depending on what went wrong and in which context). One
important information to further diagnose problems is the file in which
the problem occurred.
This change catches any exceptions and, in addition to their information
and backtrace, prints out the path of the file which caused the problem.