Add support for requiring existence of a copyright header
Not all source code files might contain copyright headers. Some of them
accidentally so.
The user might want to detect such missing headers and so this change
causes the git commit hook to assert that such a copyright header does
exist and if that is not the case, to raise an error. It furthermore
introduces a new git config option, "copyright.copyright-required" that
can be used to disable this check by setting it to false.

Import subrepo ./:copyright at eca2d00cdac7d0672f60a039084f410b882b496f