Introduce non-modifying actions
Until now the git pre-commit hook unconditionally tried to directly
fixup any problems in the to-be-committed files in-place. This is a
reasonable approach in a lot of cases but there might be others where a
less intrusive technique is desired.
This change introduces three different 'actions' that decide how to
react upon a required normalization. The 'fixup' action (the default)
modifies the files in-place. Using 'check' an error is raised causing
the commit to fail in case a file is detected to contain not normalized
copyright years. Lastly, 'warn' causes the hook to emit a warning while
letting the commit proceed.
As for the policies itself, the actions can be configured using git's
configuration infrastructure. The key to use is "copyright.action".