Add file change reversion detection
A workflow occurring occasionally includes the reversal of all changes
to a particular file. Usually, that is done in the form of an amended
commit. The problem with this case is that copyright header
normalization will take place nevertheless because the file is detected
as being changed.
This change introduces functionality to detect such content reversions
and, in such a case, causes the git pre-commit hook to automatically
exclude the file from normalization. Unfortunately, there appears to be
no way of knowing whether or not we are dealing with the amendment of a
commit or a new one. To that end, we have to perform this check every
time the hook is run.