Allow passing in of year to extend copyright range with
With recent changes in the copyright module, the normalization functions
accept a 'year' parameter specifying the year to extend a copyright
range with.
This patch adjusts the pre-commit hook to pass in the current year in
order to preserve the previous behavior.

Import subrepo ./:copyright at 137f8678298248fcb3b8def45d7bf36b94c1ac09