Add script to normalize copyright years in current commit
We have infrastructure for having any new commits automatically include
the current year in the copyright header. However, what sometimes is
needed when tidying up a program or library written in former years, is
the ability to account commits made in specific years in the copyright
header.
This change introduces a script that extracts the year of the current
git commit (i.e., the HEAD) and extends the copyright headers of all
files touched by this commit with this very year. Such functionality
comes in handy when used during interactive rebase in conjunction with
the --exec option.