Hook up support for ignoring certain copyright patterns
The copyright package we use recently got introduced support for
ignoring a copyright headers matching user configurable patterns. This
functionality can be useful when working on code bases with
contributions by several parties, each of which holding portions of
copyright. In such a case only the own copyright header should be
adjusted because of changes made.
This change hooks up this functionality in the git commit hook. As
for other options we use the git configuration (local or global) to
configure the patterns we want to ignore. For that we introduce a new
config key, "ignore", that can be set multiple times (typically using
git config's --add) option.