Update README.md to reflect better way to install the hook
So far we have relied on git templates for installing the pre-commit
hook. Although feasible, this turns out not to be the best way in terms
of flexibility of configuration. In particular, since the template gets
copied into each and every repository, changes to the original template
are not reflected unless each repository is reinitialized with the new
template.
With this change we update the README.md file and recommend a different
approach instead: the usage of the core.hooksPath git configuration
value. The hooks installed at the path pointed at by this variable are
checked directly, whenever a hook is to be invoked. This way, we retain
a single point of configuration.