Allow for easy execution of individual tests
Occasionally running the entire test suite is too much overhead or takes
too long, for instance, while working on an individual test. In such
cases, what is desired ideally is to be able to run a single test case
by means of a single memorable command.
This change adjusts our makefile to support such cases by treating every
command that is not 'test' as the individual test case to execute. Right
now, since we only have one actual TestCase object we assume the
argument supplied represents the name of a method in this class.