Correctly handle submodules
Submodules pose a problem to the copyright hook in that they make the
hook fail. In particular, the committed file representing the submodule
is actually a directory. For obvious reasons, this directory cannot
simply be opened for reading as a normal file can be and checked for a
copyright header.
To that end, this change adds a filtering step that removes all
directories from the list of files to check.
Furthermore, additional submodule meta information is kept in a
.gitmodules file. This file will never contain an actual copyright
header. Since it is also unlikely that any other hidden files are meant
to contain such a header, we additionally filter out all hidden files
from the list of files to check.